KAM_Face_pipeline_demo/requirements.txt
# Core numerical and image dependencies
numpy
opencv-python
Pillow
scikit-image

# Deep learning stack
torch
torchvision
timm
ultralytics
